Split vertex attribute logic into seperate file
